// src/common/enums/operator-type.enum.ts
export enum OperatorType {
  EQUAL = 'EQUAL',
  NOT_EQUAL = 'NOT_EQUAL',
  INCLUDE = 'INCLUDE',
  EXCLUDE = 'EXCLUDE',
  LIKE = 'LIKE',
  I_LIKE = 'I_LIKE',
  GREATER_THAN = 'GREATER_THAN',
  GREATER_THAN_OR_EQUAL = 'GREATER_THAN_OR_EQUAL',
  LESS_THAN = 'LESS_THAN',
  LESS_THAN_OR_EQUAL = 'LESS_THAN_OR_EQUAL',
  BETWEEN = 'BETWEEN',
  IS_NULL = 'IS_NULL',
  IS_NOT_NULL = 'IS_NOT_NULL',
}